fix codeclimate
