+++ app/src/main/java/org/ole/planet/takeout/SyncActivity.java
@@ -146,31 +146,6 @@ abstract class SyncActivity extends AppCompatActivity {
         spinner.setAdapter(spinnerArrayAdapter);
     }
 
-    public void setUrlParts(String url, String password, Context context) {
-        this.context = context;
-        URI uri = URI.create(url);
-        String url_Scheme = uri.getScheme();
-        String url_Host = uri.getHost();
-        int url_Port = uri.getPort();
-        String url_user = null, url_pwd = null;
-        if (url.contains("@")) {
-            String[] userinfo = uri.getUserInfo().split(":");
-            url_user = userinfo[0];
-            url_pwd = userinfo[1];
-        } else {
-            url_user = "";
-            url_pwd = password;
-        }
-        SharedPreferences.Editor editor = settings.edit();
-        editor.putString("serverURL", url);
-        editor.putString("url_Scheme", url_Scheme);
-        editor.putString("url_Host", url_Host);
-        editor.putInt("url_Port", url_Port);
-        editor.putString("url_user", url_user);
-        editor.putString("url_pwd", url_pwd);
-        editor.commit();
-        syncDatabase("_users");
-    }
 
     public void syncDatabase(final String databaseName) {
         Thread td = new Thread(new Runnable() {
